Fix yml file props and rename to related_articles
