--- examples/stories/four-column-article.stories.jsx
@@ -1,8 +1,0 @@
-import FourColumnArticle from "@/components/four_column_article";
-
-export default {
-  title: "Components/Four Column Article",
-  component: FourColumnArticle,
-};
-
-export const Default = {};
