+++ examples/stories/related-articles.stories.jsx
@@ -1,0 +1,8 @@
+import RelatedArticlesContainer from "@/components/related_articles";
+
+export default {
+  title: "Components/Related Articles",
+  component: RelatedArticlesContainer,
+};
+
+export const Default = {};
